start a cleanning
