+++ packages/react/src/oidc/index.ts
@@ -3,5 +3,5 @@ export { OidcProvider } from './OidcProvider';
 export { OidcSecure, withOidcSecure } from './OidcSecure';
 export { useOidc, useOidcAccessToken, useOidcIdToken } from './ReactOidc';
 export { OidcUserStatus, useOidcUser } from './User';
-export type { OidcConfiguration } from './vanilla/oidc';
 export { TokenRenewMode } from './vanilla/parseTokens';
+export type { AuthorityConfiguration, OidcConfiguration, StringMap } from './vanilla/types';
