+++ packages/react/src/oidc/vanilla/index.ts
@@ -1,2 +1,2 @@
-export { OidcConfiguration } from './oidc';
+export { AuthorityConfiguration, OidcConfiguration, StringMap } from './types';
 export { VanillaOidc } from './vanillaOidc';
